Prevent error when using any module with a release candidate
Some modules on the forge have versions like 0.5.0-rc1. It turns out
that the - causes havoc in Librarian, resulting in an exception that
blocks the use of librarian-puppet. I saw this first with the
puppetlabs/apache module. Basically if you include it or anything that
depends on it librarian-puppet won't work.

This patch resolves this in a hacky, but apparently working, way. It
replaces - with a period internally, and then swaps it again to a - when
making a call to the forge.

Conflicts:
	features/install.feature